moved example to mongo db connection + update compose
