+++ pom.xml
@@ -75,16 +75,10 @@
       <artifactId>spring-boot-starter-actuator</artifactId>
       <version>2.6.7</version>
     </dependency>
-    <!-- jpa, crud repository -->
+
     <dependency>
       <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-data-jpa</artifactId>
-    </dependency>
-
-    <!-- PostgreSQL -->
-    <dependency>
-      <groupId>org.postgresql</groupId>
-      <artifactId>postgresql</artifactId>
+      <artifactId>spring-boot-starter-data-mongodb</artifactId>
     </dependency>
 
     <dependency>
