+++ src/main/java/it/gov/pagopa/rtd/ms/enrolledpaymentinstrument/infrastructure/persistence/jpa/EnrolledPaymentInstrumentDao.java
@@ -1,9 +1,9 @@
 package it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.infrastructure.persistence.jpa;
 
 import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.infrastructure.persistence.jpa.model.EnrolledPaymentInstrumentEntity;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EnrolledPaymentInstrumentDao extends JpaRepository<EnrolledPaymentInstrumentEntity, UUID> { }
+public interface EnrolledPaymentInstrumentDao extends
+    MongoRepository<EnrolledPaymentInstrumentEntity, String> { }
