+++ src/main/java/it/gov/pagopa/rtd/ms/enrolledpaymentinstrument/infrastructure/persistence/jpa/model/EnrolledPaymentInstrumentEntity.java
@@ -1,34 +1,23 @@
 package it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.infrastructure.persistence.jpa.model;
 
-import java.util.UUID;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.Setter;
-import lombok.ToString;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
 
-@Entity
-@ToString
+@Document("enrolled_payment_instrument")
+@Data
+@Setter(value = AccessLevel.PACKAGE)
 @Builder(toBuilder = true)
 @AllArgsConstructor(access = AccessLevel.PACKAGE)
-@Setter(value = AccessLevel.PACKAGE)
-@Data
-@Table(name = "rtd_enrolled_payment_instrument")
 public class EnrolledPaymentInstrumentEntity {
 
   @Id
-  @GeneratedValue(strategy = GenerationType.AUTO)
-  @Column(name = "id")
-  private UUID id;
+  private String id;
 
-  @Column(name = "something")
   private String something;
 
   public EnrolledPaymentInstrumentEntity() {
